# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A high-performance subscription tracking service built with Go and PostgreSQL.\nThis API handles user authentication, subscription lifecycle management, and expense analytics.",
        "title": "SubScout Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Ayan Chatterjee",
            "url": "https://github.com/Ayan25844",
            "email": "ayan25844@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Public route to log in the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Routes"
                ],
                "summary": "Create a new login",
                "parameters": [
                    {
                        "description": "New Login",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Login"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Token created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Invalid email or password",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Public route to create a user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Routes"
                ],
                "summary": "Register a new user account",
                "parameters": [
                    {
                        "description": "New Account",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User_DTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User account created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Email id already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/users/me": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "User route to update a user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Routes"
                ],
                "summary": "Update the profile of a user",
                "parameters": [
                    {
                        "description": "Updated user profile",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Update_User_Profile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile updated successfully",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Missing token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Email id already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/me/delete": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "User route to delete a user account for the logged-in user",
                "tags": [
                    "User Routes"
                ],
                "summary": "Delete a user account",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized: Missing token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/me/password": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "User route to update a user password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Routes"
                ],
                "summary": "Update the password of a user",
                "parameters": [
                    {
                        "description": "Updated user password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Update_User_Password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User password updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Missing token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Password already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Login": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "xyz@gmail.com"
                },
                "password": {
                    "type": "string",
                    "format": "password"
                }
            }
        },
...
        "Update_User_Password": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string",
                    "format": "password"
                },
                "old_password": {
                    "type": "string",
                    "format": "password"
                }
            }
        },
        "Update_User_Profile": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string",
                    "example": "INR"
                },
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "xyz@gmail.com"
                }
            }
        },
        "User": {
            "description": "Detailed information about a user's profile",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-03-09T12:00:00Z"
                },
                "currency": {
                    "type": "string",
                    "example": "INR"
                },
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "xyz@gmail.com"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-03-09T12:00:00Z"
                }
            }
        },
        "User_DTO": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "currency": {
                    "type": "string",
                    "example": "INR"
                },
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "xyz@gmail.com"
                },
                "password": {
                    "type": "string",
                    "format": "password"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT Authorization header using the Bearer scheme.\nExample: \"Authorization: Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}